fixes for osquery pull request 2853
